requirements.txt
amqp==5.0.5
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django~=3.1.1
django-admin-list-filter-dropdown==1.0.3
django-redis-cache==3.0.0
djangorestframework~=3.11.1
hiredis==1.1.0
html-sanitizer==1.9.1
idna==2.10
kombu==5.0.2
lxml==4.6.3
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
python-telegram-handler==2.2
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
vine==5.0.0
vk-api==11.9.1
wcwidth==0.2.5
logdna~=1.7.0
psycopg2~=2.8.6